Reorganizing binaries and test structure, adding the activator test.
