+++ README.md
@@ -2,11 +2,10 @@
 
 
 <center>
-    <img src="jems.png" />
-
-    Dependency Injection (Node.js)
+    <img src="jems.png" />    
 </center>
 
+#### Dependency Injection (Node.js)
 
 [![Build Status](https://travis-ci.org/JemsFramework/di.svg?branch=release-current)](https://travis-ci.org/JemsFramework/di)
 [![npm version](https://badge.fury.io/js/%40jems%2Fdi.svg)](https://badge.fury.io/js/%40jems%2Fdi)
