+++ test/ContextualActivator.Spec.ts
@@ -1,17 +1,14 @@
 /// <reference path="../typings/index.d.ts" />
 
-import assert from 'assert'
-
+import * as assert from 'assert'
 import { ContextualActivator } from '../source/ContextualActivator'
 
 describe('The [ContainerActivator]', function() {
-
     it('should return the result of the setted function with the requested name.', function() {
         let contextualActivator:ContextualActivator = new ContextualActivator();
         contextualActivator.setContextInstantiator<any, any>('settedFunction', function(contextTypeIntance, instanceIdentifier) { return 100; });
         assert.equal(contextualActivator.getContextInstantiator<any, any>('settedFunction')(null, null), 100, 'The result is not the expected one');
     })
-
 
     it('should throw an error is the requested name is not registered.', function() {
         try {
@@ -19,7 +16,7 @@ describe('The [ContainerActivator]', function() {
         contextualActivator.getContextInstantiator<any, any>('settedFunction')(null, null);
         }
         catch(error) {
-            assert.ifError(error);
+            assert.ok(error);
         }
     })
 })
