+++ test/TestAliasSufixingConfiguration.ts
@@ -1,8 +1,8 @@
 /// <reference path="../typings/index.d.ts" />
 
 import * as assert from 'assert'
-import * as jemsdi from "../distribution/Index";
-import { IContainer } from "../distribution/IContainer";
+import * as jemsdi from "../source/Index";
+import { IContainer } from "../source/IContainer";
 
 describe('with an alias that contain sufixing configuration', function() {
