+++ test/TestContentionFluentSupport.ts
@@ -1,11 +1,11 @@
 /// <reference path="../typings/index.d.ts" />
 
 import * as assert from 'assert'
-import * as jemsdi from "../distribution/Index";
+import * as jemsdi from "../source/Index";
 import { FakeTypeA } from './fake_types/FakeTypeA';
 import { FakeTypeB } from './fake_types/FakeTypeB';
 import { FakeTypeC } from "./fake_types/FakeTypeC";
-import { IContainer } from "../distribution/IContainer";
+import { IContainer } from "../source/IContainer";
 
 describe('with fluent supported containeraized resolution', function() {
