+++ test/TestErrorTriggering.ts
@@ -1,8 +1,8 @@
 /// <reference path="../typings/index.d.ts" />
 
 import * as assert from 'assert'
-import * as jemsdi from "../distribution/Index";
-import { IContainer } from "../distribution/IContainer";
+import * as jemsdi from "../source/Index";
+import { IContainer } from "../source/IContainer";
 import { FakeTypeDependant1 } from "./fake_types/FakeTypeDependant1";
 import { FakeTypeDependant2 } from "./fake_types/FakeTypeDependant2";
 import { FakeTypeC } from "./fake_types/FakeTypeC";
